Create Q1.java
Write a Java program to display the character rhombus structure. You can assume a character to create rhombus.  
Definition of Done: 
DoD1: The program asks the user for the number of rows for the rhombus structure from the user 
DoD 2: After a successful run, the program should ask the user if he wants to try the program for some other value, and should exit only when the user wants to exit (by pressing the character ‘N’).