Create Q5.java
Create your own exception class using the extends keyword. Write a constructor for this class that takes a String argument and stores it inside the object with a String reference. Write a method that prints out the stored String. Create a try-catch clause to exercise your new exception.

Definition of Done:
1.  	Accept a string argument
2.  	Create a separate function to print the string.