Create Q5.java
Compute the sum of the series: 1+x+x2 /2!+x3/3!+…The program should ask the user to enter the value of x. The program should ask the user to enter the number of terms for which the sum is to be calculated.
Definition of Done:
a) The program should ask the user to enter the value of x.
b) The program should ask the user to enter the number of terms for which the sum is to be calculated.