Create Q6,java
Write a Java program that reads in two floating-point numbers and tests   whether they are the same up to three decimal places. 
Definition of Done:
a) The program should read two float numbers.
b) The program should ask the user to re-enter of the values entered are not numbers.
c) The program should display an appropriate message after comparison of the two numbers.