Create Q5.java
Write an application that inputs one number consisting of five digits from the user, separates the number into its individual digits in the same order and prints the digits separated from one another by three spaces each. 
Definition of Done:
a) The program should ask the user to input the integer.
b) If the number of digits in the integer is not equal to 5, the program should generate an error message and ask the use to re-enter the integer.
c) After a successful run, the program should ask the user if he wants to try the program for some other value, and should exit only when the user wants to exit (by pressing the character ‘N’).