Adding Two Distances
Write a program to read two distances in feet and inches and find their sum.  

Definition of Done: 
a) Two java files to be defined. One for class definitions and another for the application 
b) A class distance is defined with two float variables for feet and inch and the following functionality: 
i. Non-parametrized and parameterized constructors are defined. 
ii. Get and set methods are defined for all the instance variables. 
iii. Sum function is defined to calculate the sum of the two distance variables. 
iv. Display function is defined with width of 7 and precision of